+++ automation/__init__.py
@@ -1,0 +1,4 @@
+from automation.automation import Automation
+from log import log
+
+auto = Automation("MuMu模拟器12", log)
