+++ config/__init__.py
@@ -1,10 +1,9 @@
-import os
-from core.config.config import Config
+from config.config import Config
 # from pylnk3 import Lnk
 
 
-VERSION_PATH = "./config/version.txt"
-EXAMPLE_PATH = "./config/config.example.yaml"
+VERSION_PATH = "./res/config/version.txt"
+EXAMPLE_PATH = "./res/config/config.example.yaml"
 CONFIG_PATH = "./config.yaml"
 
 cfg = Config(VERSION_PATH, EXAMPLE_PATH, CONFIG_PATH)
