--- core/automation/__init__.py
@@ -1,4 +1,0 @@
-from core.automation.automation import Automation
-from core.log import log
-
-auto = Automation("MuMu模拟器12", log)
