+++ launcher/__init__.py
@@ -1,9 +1,9 @@
 import sys
 import time
-from core.config import cfg
-from core.log import log
+from config import cfg
+from log import log
 from .game_launcher import GameLauncher
-from core.automation import auto
+from automation import auto
 from utils.time_utils import TimeUtil
 
 launch = GameLauncher(cfg.game_path, cfg.game_process_name, cfg.game_type, cfg.window_name, cfg.window_class, log)
