+++ log/colorcodefilter.py
