+++ login/__init__.py
@@ -1,7 +1,7 @@
 import time
-from core.config import cfg
-from core.log import log
-from core.automation import auto
+from config import cfg
+from log import log
+from automation import auto
 from utils.time_utils import TimeUtil
