+++ res/config/reward_screenshot.json
